Move all tbi_utils.py functions into tbi.py
