+++ taxcalc/tbi/__init__.py
@@ -1,3 +1,11 @@
-from taxcalc.tbi.tbi import (run_nth_year_taxcalc_model,
+from taxcalc.tbi.tbi import (reform_warnings_errors,
+                             create_dict_table,
+                             run_nth_year_taxcalc_model,
                              run_nth_year_gdp_elast_model,
-                             reform_warnings_errors)
+                             check_years, check_user_mods,
+                             calculator_objects,
+                             random_seed, fuzzed,
+                             summary_aggregate,
+                             summary_dist_xbin, summary_diff_xbin,
+                             summary_dist_xdec, summary_diff_xdec,
+                             check_years_return_first_year)
